refactor organize components to atomic design folders
