+++ components/moleculs/modal/modal.styled.ts
@@ -1,0 +1,23 @@
+import styled from "styled-components"
+
+export const Background = styled.section`
+width:100%;
+height:100vh;
+display: flex;
+justify-content: center;
+align-items: center;
+background-color: rgba(0,0,0,0.9);
+`
+
+export const Modal = styled.div`
+width:34vw;
+height: 25vh;
+min-height: 350px;
+display: flex;
+flex-direction: column;
+justify-content: center;
+align-items: center;
+border-radius: 10px;
+background-color: #212222;
+padding:2rem;
+`
