+++ components/templates/layout-center/index.ts
