+++ pages/auth/signin.tsx
@@ -8,7 +8,7 @@ import googleLogo from "../../public/icons/google-logo.jpeg";
 import kakaoLogo from "../../public/icons/kakao-logo.png";
 import naverLogo from "../../public/icons/naver-logo.png";
 import * as S from "./signin.styled";
-import {LayoutCenter} from "components/layout-center";
+import {LayoutCenter} from "components/templates/layout-center";
 
 export default function SignIn({
   providers,
